PullRequest: 12 更新 cu_seqlens 数据类型为 int32，保持和 vllm/sglang一致
Merge branch 'fix/cu-seqlens-int32 of git@code.alipay.com:ling/flashla.git into opensource/v0.1

https://code.alipay.com/ling/flashla/pull_requests/12




* fix: update cu_seqlens data type from int64 to int32(consistent with vllm/sglang)
* fix: keep .gitignore clean

+++ benchmarks/bench_kda_fused_fwd.py
@@ -139,7 +139,7 @@ def bench_fixed(configs):
         torch.cuda.empty_cache()
 
         seq_lens = [T] * B
-        cu_seqlens = torch.tensor(exclusive_cumsum(seq_lens), dtype=torch.int64, device=device)
+        cu_seqlens = torch.tensor(exclusive_cumsum(seq_lens), dtype=torch.int32, device=device)
 
         inputs = prepare_safe_gate_inputs(B, T, H, D, device, cu_seqlens=cu_seqlens)
         q, k, v, g, beta = inputs['q'], inputs['k'], inputs['v'], inputs['g'], inputs['beta']
@@ -189,7 +189,7 @@ def bench_varlen(configs):
         torch.cuda.empty_cache()
 
         T = total_len
-        cu_seqlens = torch.tensor(exclusive_cumsum(seq_lens), dtype=torch.int64, device=device)
+        cu_seqlens = torch.tensor(exclusive_cumsum(seq_lens), dtype=torch.int32, device=device)
 
         inputs = prepare_safe_gate_inputs(1, T, H, D, device, cu_seqlens=cu_seqlens)
         q, k, v, g, beta = inputs['q'], inputs['k'], inputs['v'], inputs['g'], inputs['beta']
