+++ csrc/kda/sm90/kda_fwd_sm90_safe_gate.cu
@@ -14,7 +14,7 @@ template void launch_kda_fwd_prefill_kernel_gbai<
     false, true, true, false, true,
     cutlass::arch::Sm90, bf16, bf16, float>(
     cudaStream_t, bf16*, float*, bf16 const*, bf16 const*, bf16 const*,
-    float const*, float const*, float const*, int64_t const*, uint8_t*,
+    float const*, float const*, float const*, int32_t const*, uint8_t*,
     int32_t, int32_t, int32_t, int32_t, int32_t, int32_t, int64_t, float, int32_t);
 
 // SafeGate=true, InitState=true
@@ -22,7 +22,7 @@ template void launch_kda_fwd_prefill_kernel_gbai<
     false, true, true, true, true,
     cutlass::arch::Sm90, bf16, bf16, float>(
     cudaStream_t, bf16*, float*, bf16 const*, bf16 const*, bf16 const*,
-    float const*, float const*, float const*, int64_t const*, uint8_t*,
+    float const*, float const*, float const*, int32_t const*, uint8_t*,
     int32_t, int32_t, int32_t, int32_t, int32_t, int32_t, int64_t, float, int32_t);
 
 }  // namespace flat
